backend/requirements.txt
aiogram==3.4.1
fastapi==0.110.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.29
gspread==6.1.0
oauth2client==4.1.3
python-dotenv==1.0.1
pydantic==2.5.3
git+https://github.com/nimaxin/init-data-py.git
apscheduler==3.10.4